Error_Exception: Extend to support serialization.
Make the SimpleSAML_Error_Exception class suitable for serialization by
making it _not_ serialize the trace variable in the Exception class.

Serializing the trace-variable is problematic because it contains the
parameters to all function calls. This can make the exception very big.
This can also make the exception unserializable, if one of the
parameters is unserializable.